delete old commit plugin
