+++ commitrmnocb/validate_observation.go
@@ -1,9 +1,11 @@
 package commitrmnocb
 
 import (
+	"context"
 	"fmt"
 
 	mapset "github.com/deckarep/golang-set/v2"
+	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
 	cciptypes "github.com/smartcontractkit/chainlink-common/pkg/types/ccipocr3"
 
@@ -11,7 +13,7 @@ import (
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/ocr3types"
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 
-	"github.com/smartcontractkit/chainlink-ccip/commit"
+	"github.com/smartcontractkit/chainlink-ccip/internal/reader"
 	"github.com/smartcontractkit/chainlink-ccip/plugintypes"
 )
 
@@ -48,11 +50,11 @@ func (p *Plugin) ValidateObservation(_ ocr3types.OutcomeContext, _ types.Query, 
 		return fmt.Errorf("failed to validate OffRampNextSeqNums: %w", err)
 	}
 
-	if err := commit.ValidateObservedTokenPrices(obs.TokenPrices); err != nil {
+	if err := validateObservedTokenPrices(obs.TokenPrices); err != nil {
 		return fmt.Errorf("failed to validate token prices: %w", err)
 	}
 
-	if err := commit.ValidateObservedGasPrices(obs.GasPrices); err != nil {
+	if err := validateObservedGasPrices(obs.GasPrices); err != nil {
 		return fmt.Errorf("failed to validate gas prices: %w", err)
 	}
 
@@ -143,4 +145,79 @@ func validateFChain(fChain map[cciptypes.ChainSelector]int) error {
 	}
 
 	return nil
+}
+
+func validateObservedTokenPrices(tokenPrices []cciptypes.TokenPrice) error {
+	tokensWithPrice := mapset.NewSet[types.Account]()
+	for _, t := range tokenPrices {
+		if tokensWithPrice.Contains(t.TokenID) {
+			return fmt.Errorf("duplicate token price for token: %s", t.TokenID)
+		}
+		tokensWithPrice.Add(t.TokenID)
+
+		if t.Price.IsEmpty() {
+			return fmt.Errorf("token price must not be empty")
+		}
+	}
+
+	return nil
+}
+
+func validateObservedGasPrices(gasPrices []cciptypes.GasPriceChain) error {
+	// Duplicate gas prices must not appear for the same chain and must not be empty.
+	gasPriceChains := mapset.NewSet[cciptypes.ChainSelector]()
+	for _, g := range gasPrices {
+		if gasPriceChains.Contains(g.ChainSel) {
+			return fmt.Errorf("duplicate gas price for chain %d", g.ChainSel)
+		}
+		gasPriceChains.Add(g.ChainSel)
+		if g.GasPrice.IsEmpty() {
+			return fmt.Errorf("gas price must not be empty")
+		}
+	}
+
+	return nil
+}
+
+// validateMerkleRootsState merkle roots seq nums validation by comparing with on-chain state.
+func validateMerkleRootsState(
+	ctx context.Context,
+	lggr logger.Logger,
+	report cciptypes.CommitPluginReport,
+	reader reader.CCIP,
+) (bool, error) {
+	reportChains := make([]cciptypes.ChainSelector, 0)
+	reportMinSeqNums := make(map[cciptypes.ChainSelector]cciptypes.SeqNum)
+	for _, mr := range report.MerkleRoots {
+		reportChains = append(reportChains, mr.ChainSel)
+		reportMinSeqNums[mr.ChainSel] = mr.SeqNumsRange.Start()
+	}
+
+	if len(reportChains) == 0 {
+		return true, nil
+	}
+
+	onchainNextSeqNums, err := reader.NextSeqNum(ctx, reportChains)
+	if err != nil {
+		return false, fmt.Errorf("get next sequence numbers: %w", err)
+	}
+	if len(onchainNextSeqNums) != len(reportChains) {
+		return false, fmt.Errorf("critical error: onchainSeqNums length mismatch")
+	}
+
+	for i, nextSeqNum := range onchainNextSeqNums {
+		chain := reportChains[i]
+		reportMinSeqNum, ok := reportMinSeqNums[chain]
+		if !ok {
+			return false, fmt.Errorf("critical error: reportSeqNum not found for chain %d", chain)
+		}
+
+		if reportMinSeqNum != nextSeqNum {
+			lggr.Warnw("report is not valid due to seq num mismatch",
+				"chain", chain, "reportMinSeqNum", reportMinSeqNum, "onchainNextSeqNum", nextSeqNum)
+			return false, nil
+		}
+	}
+
+	return true, nil
 }
